requirements.txt
arrow==1.2.2
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2022.5.18.1
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==2.1.1
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
djangorestframework==3.13.1
filelock==3.7.1
identify==2.5.1
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
nodeenv==1.6.0
oauthlib==3.2.0
Pillow==9.1.1
platformdirs==2.5.2
pre-commit==2.19.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.9
virtualenv==20.14.1
